refactor : calculator 파일 구조화
