+++ src/js/calculator/view.js
